pyfem/requirements.txt
Flask
Flask-PyMongo
Flask-DebugToolbar
Flask-YAMLConfig
Jinja2
WTForms
Werkzeug
mongoengine
#pymongo
flask-mongoengine
python-dateutil
#Flask-Script
path.py
pyyaml
markdown2
translitcodec

# Elastic Search
#pyes


# deployment
fabric

# testing
#unittest2
#pytest
#pytest-pep8
#pep8

# docs
#Pygments==1.4
#Sphinx==1.0.7
